+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - [NEW] Particle emitter component in Scene
 - [NEW] Lit particles in ParticleRenderer
 - [NEW] ResourcePath class to unify file access
+- [NEW] Shader #import directive